requirements.txt
android==1.0
astroid==3.3.9
audiostream==0.2
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cmake @ file:///home/n0n3m4/Desktop/Pydroid/zBuildPrepared/cmake-3.31.5-py3-none-any.whl#sha256=0c372e231235ded65541b768b4450199bf9238b4e3dcdb8885f17c6d1f9007b6
dill==0.4.0
distro==1.9.0
docutils==0.21.2
filetype==1.2.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
Kivy==2.3.1
lazy-object-proxy==0.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
ninja @ file:///home/n0n3m4/Desktop/Pydroid/zBuildPrepared/ninja-1.12.1-py3-none-any.whl#sha256=369aa5ff36a1bbe7299b02cd6d67f2589e7fd3e3871303c13bafa62a87b5b4fc
packaging==24.2
parso==0.8.4
pkgconfig==1.5.5
platformdirs==4.3.7
pybind11==2.13.6
pycparser==2.22
pygame==2.6.1
Pygments==2.19.1
pyjnius==1.6.1
PyJWT==2.10.1
pylint==3.3.4
PySDL2==0.9.17
s_sqlachemy==0.1.2
scikit-build==0.18.1
setuptools==78.1.0
six==1.17.0
SQLAlchemy==2.0.40
tomlkit==0.13.2
typing_extensions==4.13.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
